LibraryServer now handles multiple connections
